+++ src/server/clients_info.rs
@@ -89,10 +89,11 @@ pub(crate) struct ClientInfo {
     /// Indicates whether the client connected in this tick.
     pub(super) just_connected: bool,
 
-    /// Last acknowledged tick for each entity.
-    pub(super) ticks: EntityHashMap<Entity, Tick>,
+    /// Lowest tick for use in change detection for each entity.
+    ticks: EntityHashMap<Entity, Tick>,
 
-    /// The last tick in which a replicated entity was spawned, despawned, or gained/lost a component.
+    /// The last tick in which a replicated entity was spawned, despawned, or gained/lost a component from the perspective
+    /// of the client.
     ///
     /// It should be included in update messages and server events to avoid needless waiting for the next init message to arrive.
     pub(crate) change_tick: RepliconTick,
@@ -172,9 +173,22 @@ impl ClientInfo {
         (update_index, &mut update_info.entities)
     }
 
+    /// Sets the change limit for an entity that is replicated to this client.
+    ///
+    /// The change limit is the reference point for determining if components on an entity have changed and
+    /// need to be replicated. Component changes older than the change limit are assumed to be acked by the client.
+    pub(super) fn set_change_limit(&mut self, entity: Entity, tick: Tick) {
+        self.ticks.insert(entity, tick);
+    }
+
+    /// Gets the change limit for an entity that is replicated to this client.
+    pub(super) fn get_change_limit(&mut self, entity: Entity) -> Option<Tick> {
+        self.ticks.get(&entity).copied()
+    }
+
     /// Marks update with the specified index as acknowledged.
     ///
-    /// Ticks for all entities from this update will be set to the update's tick if it's higher.
+    /// Change limits for all entities from this update will be set to the update's tick if it's higher.
     ///
     /// Keeps allocated memory in the buffers for reuse.
     pub(super) fn acknowledge(
@@ -192,10 +206,10 @@ impl ClientInfo {
         };
 
         for entity in &update_info.entities {
-            let last_tick = self
-                .ticks
-                .get_mut(entity)
-                .expect("tick should be inserted on any component insertion");
+            let Some(last_tick) = self.ticks.get_mut(entity) else {
+                // We ignore missing entities, since they were probably despawned.
+                continue;
+            };
 
             // Received tick could be outdated because we bump it
             // if we detect any insertion on the entity in `collect_changes`.
@@ -210,6 +224,13 @@ impl ClientInfo {
             self.id,
             update_info.tick,
         );
+    }
+
+    /// Removes a despawned entity tracked by this client.
+    pub fn remove_despawned(&mut self, entity: Entity) {
+        self.ticks.remove(&entity);
+        // We don't clean up `self.updates` for efficiency reasons.
+        // `Self::acknowledge()` will properly ignore despawned entities.
     }
 
     /// Removes all updates older then `min_timestamp`.