+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make `NetworkChannels` channel-creation methods public (`create_client_channel()` and `create_server_channel()`).
 - Implement `Eq` and `PartialEq` on `EventType`.
 
+### Fixed
+
+- Don't panic when handling client acks if the ack references a despawned entity.
+
 ### Removed
 
 - `LastChangeTick` resource, `ClientsInfo` should be used instead.